Compute count probabilities using dePril convolution (bi)

Compute count probabilities using dePril convolution (section 3.2) for the built-in distributions

Compute count probabilities using dePril convolution (section 3.2) for the user passed survival function.

dCount_dePril_bi(x, distPars, dist, nsteps = 100L, time = 1,
  extrap = TRUE, cdfout = FALSE, logFlag = FALSE)

dCount_dePril_user(x, distPars, extrapolPars, survR, nsteps = 100L,
  time = 1, extrap = TRUE, cdfout = FALSE, logFlag = FALSE)

Arguments

survR
Rcpp::Function user passed survival function; should have the signature function(t, distPars) where t is a real number (>0) where the survival function is evaluated and distPars is a list of distribution parameters. It should return a double value.

inheritParams
dCount_naive_bi

Value

vector of probabilities P(x(i)) for i = 1, ..., n where n is length of x.

Details

The routine does minimum number of convolution using dePril trick to compute the count probability P(x) sing nsteps steps, and refines result by Richardson extrapolation if extrap is TRUE using the algorithm of section 3.2.