Compute count probabilities using naive convolution (bi)

Compute count probabilities using naive convolution (section 3.1) for the built-in distributions

Compute count probabilities using naive convolution (section 3.1) for the user passed survival function.

dCount_naive_bi(
  x,
  distPars,
  dist,
  nsteps = 100L,
  time = 1,
  extrap = TRUE,
  cdfout = FALSE,
  logFlag = FALSE
)

dCount_naive_user(
  x,
  distPars,
  extrapolPars,
  survR,
  nsteps = 100L,
  time = 1,
  extrap = TRUE,
  cdfout = FALSE,
  logFlag = FALSE
)

Arguments

x

integer (vector), the desired count values.

distPars

Rcpp::List with distribution specific slots, see details.

dist

character name of the built-in distribution, see details.

nsteps

unsiged integer number of steps used to compute the integral.

time

double time at wich to compute the probabilities. Set to 1 by default.

extrap

logical if TRUE, Richardson extrapolation will be applied to improve accuracy.

cdfout

logical if TRUE, the cdf will be returned instead of the count probability.

logFlag

logical if TRUE the log-probability will be returned.

extrapolPars

ma::vec of length 2. The extrapolation values.

survR

Rcpp::Function user passed survival function; should have the signature function(t, distPars) where t is a real number (>0) where the survival function is evaluated and distPars is a list of distribution parameters. It should return a double value.

Value

vector of probabilities P(x(i)) for i = 1, ..., n where n is length of x.

Details

The routine does minimum number of convolution to compute the count probability P(x) sing nsteps steps, and refines result by Richardson extrapolation if extrap is TRUE using the algorithm of section 3.1.